Use the new way to create a public S3 bucket for the pw-manager ui
This is in response to AWS's change that, by default, blocks public
access and disables ACLs. This commit follows the minimum steps
necessary (as far as I can tell) to continue to set up our public ui
bucket to have the same permissions we were giving it before.

The "ObjectWriter" ownership value was chosen because that's what our
existing IDP pw-manager ui buckets currently have on S3.

I expect us to remove all of these ui-bucket related terraform resources
when we move to using Cloudflare for the pw-manager ui.